+++ packages/expo-modules-autolinking/src/types.ts
@@ -76,6 +76,7 @@ export interface ModuleDescriptorIos extends CommonNativeModuleDescriptor {
   flags: Record<string, any> | undefined;
   swiftModuleNames: string[];
   appDelegateSubscribers: string[];
+  sceneDelegateSubscribers: string[];
   reactDelegateHandlers: string[];
   debugOnly: boolean;
 }
@@ -185,6 +186,11 @@ export type RawModuleConfigApple = {
    * Names of Swift classes that hooks into `ExpoAppDelegate` to receive AppDelegate life-cycle events.
    */
   appDelegateSubscribers?: string[];
+
+  /**
+   * Names of Swift classes that hook into `ExpoAppSceneDelegate` to receive scene life-cycle events.
+   */
+  sceneDelegateSubscribers?: string[];
 
   /**
    * Names of Swift classes that implement `ExpoReactDelegateHandler` to hook React instance creation.